+++ src/pages/Gateway/index.tsx
@@ -29,6 +29,7 @@ import {
   WRITE_OPTIONS,
   log,
 } from '@src/constants';
+import useBalances from '@src/hooks/useBalances';
 import useGateway from '@src/hooks/useGateway';
 import useGateways from '@src/hooks/useGateways';
 import useHealthcheck from '@src/hooks/useHealthCheck';
@@ -61,10 +62,10 @@ const Gateway = () => {
 
   const walletAddress = useGlobalState((state) => state.walletAddress);
   const arIOWriteableSDK = useGlobalState((state) => state.arIOWriteableSDK);
-  const balances = useGlobalState((state) => state.balances);
   const ticker = useGlobalState((state) => state.ticker);
   const { data: protocolBalance } = useProtocolBalance();
   const { data: gateways } = useGateways();
+  const { data: balances } = useBalances(walletAddress);
 
   const params = useParams();
 
@@ -400,7 +401,7 @@ const Gateway = () => {
       <div className="my-6 flex gap-6">
         <div className="flex min-w-72 flex-col gap-6">
           <div className="size-fit w-full rounded-xl border border-transparent-100-16 text-sm">
-            <div className="px-6 py-4">
+            <div className="bg-containerL3 px-6 py-4">
               <div className="text-high">Stats</div>
             </div>
             <StatsBox
@@ -458,22 +459,24 @@ const Gateway = () => {
                   }
                 />
               </>
-            ) : gateway && (
-              <StatsBox
-                title="Leave Date"
-                value={
-                  gateway?.endTimestamp
-                    ? formatDateTime(new Date(gateway?.endTimestamp))
-                    : undefined
-                }
-              />
+            ) : (
+              gateway && (
+                <StatsBox
+                  title="Leave Date"
+                  value={
+                    gateway?.endTimestamp
+                      ? formatDateTime(new Date(gateway?.endTimestamp))
+                      : undefined
+                  }
+                />
+              )
             )}
             {/* <StatsBox title="Rewards Distributed" value={gateway?} /> */}
           </div>
 
           {gateway?.weights && gateway?.status === 'joined' && (
             <div className="w-full rounded-xl border border-transparent-100-16 text-sm">
-              <div className="px-6 py-4">
+              <div className="bg-containerL3 px-6 py-4">
                 <div className="text-high">Weights</div>
               </div>
 
@@ -498,7 +501,7 @@ const Gateway = () => {
         </div>
         <div className="flex w-full grow flex-col gap-6">
           <div className="h-fit w-full overflow-hidden rounded-xl border border-transparent-100-16">
-            <div className="flex items-center py-4 pl-6 pr-3">
+            <div className="flex items-center bg-containerL3 py-4 pl-6 pr-3">
               <div className="text-sm text-high">General Information</div>
               <div className="flex grow gap-6" />
               {ownerId === walletAddress?.toString() &&