+++ src/pages/Staking/DelegateStakeTable.tsx
@@ -261,7 +261,7 @@ const DelegateStake = () => {
 
   return (
     <div>
-      <div className="flex w-full items-center rounded-t-xl border border-grey-600 py-4 pl-6 pr-3">
+      <div className="flex w-full items-center rounded-t-xl border border-grey-600 bg-containerL3 py-[0.9375rem] pl-6 pr-3">
         <div className="grow text-sm text-mid">Delegate Stake</div>
       </div>
       <TableView